finish

+++ app/dao/movie.py
@@ -5,23 +5,29 @@ class MovieDAO:
     def __init__(self, session):
         self.session = session
 
-    def get_one(self, mid):
+    def get(self, mid, **kwargs):
+        """Метод, который выводит все фильмы или фильмы по режиссёрам, жанрам и годам или фильм по id"""
+        query = self.session.query(Movie)
 
-        return self.session.query(Movie).get(mid)
+        if mid:
+            return query.get(mid)
+        if kwargs:
+            for key, value in kwargs.items():
+                query = query.filter(eval(f"Movie.{key}") == int(value))
 
-    def get_all(self):
-
-        return self.session.query(Movie).all()
+        return query.all()
 
     def create(self, data):
+        """Метод, который добавляет новый фильм"""
         movie = Movie(**data)
 
-        self.session.add(movie)
-        self.session.commit()
+        with self.session.begin():
+            self.session.add(movie)
 
         return movie
 
     def update(self, movie):
+        """Метод, который обновляет полностью или частично данные фильма"""
 
         self.session.add(movie)
         self.session.commit()
@@ -29,7 +35,9 @@ class MovieDAO:
         return movie
 
     def delete(self, mid):
-        movie = self.get_one(mid)
-
+        """Метод, который удаляет данные фильма"""
+        movie = self.get(mid)
+        if not movie:
+            return
         self.session.delete(movie)
         self.session.commit()
